Update README.md
This is a Java-based calculator application that allows users to perform basic arithmetic operations through a [command-line interface/graphical user interface/etc.]. Users can input an operation symbol (+, -, *, /) and the corresponding numbers to perform the desired calculation.

Examples:
- Input: 2 + 3
  Output: 5
- Input: 10 / 2
  Output: 5

If the user enters an invalid operation symbol or an incorrect number format, the program will display an error message: "Wrong Input. Try again" and prompt the user to re-enter the input.